requirements.txt
groq>=0.12.0
anthropic>=0.49.0
PyGithub>=2.5.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
yt-dlp>=2024.0.0
curl_cffi>=0.7.0
python-telegram-bot>=21.0
